fix(modularity): resolve L6 circular dependencies by relocating modules
Breaking the config→adapters→utils→config cycle:
1. Move diagnostics.py: utils → exceptions (breaks exceptions→utils→exceptions)
2. Move runtime_assets.py: utils → clients (breaks adapters→utils→config)
3. Use lazy imports in comment_utils for config dependencies

After fix:
- test_no_circular_deps_outside_l3_core: PASSED ✓
- All L6 circular dependencies resolved
- Modularity test suite: 10 passed, 6 xfailed, 0 xpassed

Files changed:
- src/vibe3/utils/diagnostics.py → src/vibe3/exceptions/diagnostics.py
- src/vibe3/utils/runtime_assets.py → src/vibe3/clients/runtime_assets.py
- Updated all imports (13 files total)
