+++ src/vibe3/adapters/github_flow.py
@@ -1,8 +1,8 @@
 """GitHub Flow adapter — lightweight skill distribution from global runtime."""
 
 from vibe3.adapters import register_adapter
+from vibe3.clients.runtime_assets import runtime_assets_root
 from vibe3.models.adapter_manifest import AdapterManifest, AdapterResource
-from vibe3.utils.runtime_assets import runtime_assets_root
 
 
 def _build_github_flow_manifest() -> AdapterManifest:
