+++ src/vibe3/adapters/vibe_center.py
@@ -82,7 +82,7 @@ def _build_vibe_center_manifest() -> AdapterManifest:
     )
 
     # Skills (scan directory)
-    from vibe3.utils.runtime_assets import runtime_assets_root
+    from vibe3.clients.runtime_assets import runtime_assets_root
 
     skills_dirs = [repo_root / "skills"]
     global_skills = runtime_assets_root() / "skills"
