+++ src/vibe3/clients/runtime_assets.py
