+++ src/vibe3/exceptions/diagnostics.py
